requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.1
flask>=3.0.2
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
aiohttp>=3.9.3